+++ internal/sdkprovider/service/account/account_team_member.go
@@ -110,15 +110,9 @@ func resourceAccountTeamMemberRead(ctx context.Context, d *schema.ResourceData, 
 	for _, invite := range resp {
 		if invite.UserEmail == userEmail {
 			if err = schemautil.ResourceDataSet(
-				aivenAccountTeamMemberSchema,
-				d,
-				invite,
-				schemautil.RenameAliases(map[string]string{}),
+				d, invite, aivenAccountTeamMemberSchema,
+				schemautil.SetForceNew("account_id", accountID),
 			); err != nil {
-				return err
-			}
-
-			if err = d.Set("account_id", accountID); err != nil {
 				return err
 			}
 
@@ -139,14 +133,9 @@ func resourceAccountTeamMemberRead(ctx context.Context, d *schema.ResourceData, 
 	for _, member := range respTI {
 		if member.UserEmail == userEmail {
 			if err = schemautil.ResourceDataSet(
-				aivenAccountTeamMemberSchema,
-				d,
-				member,
+				d, member, aivenAccountTeamMemberSchema,
+				schemautil.SetForceNew("account_id", accountID),
 			); err != nil {
-				return err
-			}
-
-			if err = d.Set("account_id", accountID); err != nil {
 				return err
 			}